Promote use of `stringescapes {}`
Everyone seems to actually use these, so using a different example in
the manual isn't entirely helpful, plus `[` and `]` are tokens in
Snowball making them a potentially confusing choice to someone trying
to get to grips with the language.

This also means the pygments highlighter now highlights these snippets
correctly, as that assumes `stringescapes {}` is in use.