Ensure that output directory exists
Previously, the script would fail with an obscure error if the output
directory didn't exist.

Also, display the directory that the output is being generated in, to
avoid confusion.

Also, add more dependency information to README - I found that I needed
to install python-pygments on an ubuntu system to make the generation
work.