Generate stems for sample vocabularies
Hard-coding lists of stemmed forms is boring grunt work and risks
them getting out of step with the algorithms, so add a new function
to the template which takes a sample vocabulary and renders it as
a table like we previously hard-coded with stemmed forms generated
by running the vocabulary through the stemwords program.

Fortunately the conversion shows that the currently claimed outputs
are all correct.

Suggested by David Corbett.  Fixes #11.