Remove sexist "joke" from front page
Making a "joke" about not naming the algorithm after a sex worker is both shaming to sex workers ("good sense prevailed") and stops people making contributions who find such things off-putting. It also makes no contribution ("I was thinking about onions at the time so I almost named it onion, but I didn't") aside from showing that the maintainers and contributors are fine with having a sexist joke on the web site.

As such, this patch removes it, leaving only relevant parts of the attributed quote.