Simplify the "runtime use" page
It doesn't seem helpful to suggest people pull code out of the
distribution and compile it by hand, and such instructions are
likely to bitrot over time as people change the distributed code
and fail to update these instructions.

Note that "-comments" is now needed to get comments in the generated
source code.