Add simpler way to call algorithm_vocab
It can now take a number of words and two start words, from which we
pull words from the snowball-data voc.txt for the language.

Refactor all the languages for which the vocab sample is excerpted
from the vocab list (or almost so in the case of Basque) to use this.