Delete page about Pascal generator
This was merged into the distribution more than 2 years ago, and
was in the Snowball 2.0.0 release nearly 2 years ago.

I changed the page about it into a stub page which linked to the
original on the old snowball site, but at this point such a page
no longer seems useful.