Update Hungarian links
The webpage for the Hungarian stemmer has gone and I couldn't find
a replacement URL.

Looking at the wayback machine's snapshot of it, the non-boilerplate
page content was

  We provide two stemmers for Hungarian: a light stemmer and a heavy
  stemmer. The light stemmer is less invasive than the heavy stemmer
  altering fewer words. It attempts to stem nouns and adjectives. The
  heavy stemmer also stems verbs and numerals.

plus links to the old Snowball website and a CLEF 2005 paper about
the stemmer.

I have replaced the old link with a direct link to this paper.